project/firmware/producer/requirements.txt
python-dotenv
pika
pillow
opencv-python
numpy
